Rename file login-helpers.ts -> login-library.ts
Jira ticket: CAMS-401
