+++ user-interface/src/login/providers/okta/OktaSession.test.tsx
@@ -1,7 +1,7 @@
 import { describe } from 'vitest';
 import { OktaSession } from './OktaSession';
 import { render, screen, waitFor } from '@testing-library/react';
-import { CamsUser } from '@/login/login-helpers';
+import { CamsUser } from '@/login/login-library';
 import { BrowserRouter } from 'react-router-dom';
 import * as sessionModule from '../../Session';
 import * as oktaReactModule from '@okta/okta-react';
